+++ GraphicsFramework/src/graphics/VertexBuffer.h
@@ -7,7 +7,7 @@ class VertexBuffer : public GraphicsResource<VertexBuffer>
 {
 public:
 
-	VertexBuffer(size_t size, BufferUsage usage);
+	VertexBuffer(unsigned int size, BufferUsage usage);
 	virtual ~VertexBuffer();
 
 	virtual void SetData(const void* data) = 0;
@@ -15,12 +15,12 @@ public:
 	BufferUsage GetUsage() const;
 	virtual size_t GetSize() const;
 
-	static VertexBuffer* Create(size_t size, BufferUsage usage = BufferUsage::STATIC, void* data = nullptr);
+	static VertexBuffer* Create(unsigned int size, BufferUsage usage = BufferUsage::STATIC, void* data = nullptr);
 
 	bool operator==(const VertexBuffer* other) const;
 
 protected:
-	size_t m_size;
+	unsigned int m_size;
 	BufferUsage m_usage;
 
 };
